nserver: protocol.h: Update sssample.
* nserver/src/protocol.c (sssample):  Return mean on success.
* nserver/src/protocol.h (sssample): Change return data type to
double.
* nserver/tests/protocol_tests.c (test_sssample): Update test.

+++ nserver/src/protocol.h
@@ -10,7 +10,7 @@
 
 int sscreate(char *key);
 int ssdelete(char *key);
-int sssample(char *key, double s);
+double sssample(char *key, double s);
 double ssmean(char *key);
 char *ssdump(char *key);
 char *sslist();
